+++ backend/schedulerd/schedulerd.go
@@ -10,6 +10,7 @@ import (
 	"github.com/sensu/sensu-go/backend/ringv2"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/backend/store/cache"
+	"github.com/sensu/sensu-go/secrets"
 	"github.com/sensu/sensu-go/types"
 )
 
@@ -46,16 +47,17 @@ var (
 // Schedulerd handles scheduling check requests for each check's
 // configured interval and publishing to the message bus.
 type Schedulerd struct {
-	store                store.Store
-	queueGetter          types.QueueGetter
-	bus                  messaging.MessageBus
-	checkWatcher         *CheckWatcher
-	adhocRequestExecutor *AdhocRequestExecutor
-	ctx                  context.Context
-	cancel               context.CancelFunc
-	errChan              chan error
-	ringPool             *ringv2.Pool
-	entityCache          *cache.Resource
+	store                  store.Store
+	queueGetter            types.QueueGetter
+	bus                    messaging.MessageBus
+	checkWatcher           *CheckWatcher
+	adhocRequestExecutor   *AdhocRequestExecutor
+	ctx                    context.Context
+	cancel                 context.CancelFunc
+	errChan                chan error
+	ringPool               *ringv2.Pool
+	entityCache            *cache.Resource
+	secretsProviderManager *secrets.ProviderManager
 }
 
 // Option is a functional option.
@@ -63,21 +65,23 @@ type Option func(*Schedulerd) error
 
 // Config configures Schedulerd.
 type Config struct {
-	Store       store.Store
-	QueueGetter types.QueueGetter
-	RingPool    *ringv2.Pool
-	Bus         messaging.MessageBus
-	Client      *clientv3.Client
+	Store                  store.Store
+	QueueGetter            types.QueueGetter
+	RingPool               *ringv2.Pool
+	Bus                    messaging.MessageBus
+	Client                 *clientv3.Client
+	SecretsProviderManager *secrets.ProviderManager
 }
 
 // New creates a new Schedulerd.
 func New(ctx context.Context, c Config, opts ...Option) (*Schedulerd, error) {
 	s := &Schedulerd{
-		store:       c.Store,
-		queueGetter: c.QueueGetter,
-		bus:         c.Bus,
-		errChan:     make(chan error, 1),
-		ringPool:    c.RingPool,
+		store:                  c.Store,
+		queueGetter:            c.QueueGetter,
+		bus:                    c.Bus,
+		errChan:                make(chan error, 1),
+		ringPool:               c.RingPool,
+		secretsProviderManager: c.SecretsProviderManager,
 	}
 	s.ctx, s.cancel = context.WithCancel(ctx)
 	cache, err := cache.New(s.ctx, c.Client, &corev2.Entity{}, true)
@@ -85,8 +89,8 @@ func New(ctx context.Context, c Config, opts ...Option) (*Schedulerd, error) {
 		return nil, err
 	}
 	s.entityCache = cache
-	s.checkWatcher = NewCheckWatcher(s.ctx, c.Bus, c.Store, c.RingPool, cache)
-	s.adhocRequestExecutor = NewAdhocRequestExecutor(s.ctx, s.store, s.queueGetter.GetQueue(adhocQueueName), s.bus, s.entityCache)
+	s.checkWatcher = NewCheckWatcher(s.ctx, c.Bus, c.Store, c.RingPool, cache, s.secretsProviderManager)
+	s.adhocRequestExecutor = NewAdhocRequestExecutor(s.ctx, s.store, s.queueGetter.GetQueue(adhocQueueName), s.bus, s.entityCache, s.secretsProviderManager)
 
 	for _, o := range opts {
 		if err := o(s); err != nil {