+++ secrets/logger.go
@@ -1,0 +1,7 @@
+package secrets
+
+import "github.com/sirupsen/logrus"
+
+var logger = logrus.WithFields(logrus.Fields{
+	"component": "secrets",
+})