+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added the secrets provider interface and secrets provider manager to be used
+by commerical secrets providers. Implemented for checks, mutators, and handlers.
+
 ## [5.16.0] - 2019-12-11
 
 ### Added